start reorganizing atom/bond featurizers
