+++ chemprop/v2/featurizers/multihot/__init__.py
@@ -1,3 +1,3 @@
-from .atom import AtomFeaturizer
-from .bond import BondFeaturizer
+from ..atom import AtomFeaturizer
+from ..bond import BondFeaturizer
 from .base import MultiHotFeaturizer
